pyproject.toml
[project]
name = "my-example-project"
version = "1.0.0"
description = "Sagemaker + MLFlow project with separate SageMaker utilities"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "boto3>=1.26.0",
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "scikit-learn>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=0.19.0",
    "sagemaker>=2.190.0",
    "openpyxl>=3.1.5",
    "fsspec>=2025.3.0",
    "s3fs>=0.4.2",
    "seaborn>=0.13.2",
    "matplotlib>=3.7.5",
    "mlflow==2.17.2",  # MLFlow version has to be locked to ensure compatibility with SageMaker
    "xgboost>=2.1.4",
    "joblib>=1.4.2",
    "sagemaker-mlflow>=0.1.0",
    "ucimlrepo>=0.0.7",
]
requires-python = ">=3.8"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.2",
    "moto>=4.2.0",
    "docker>=6.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "raise ImportError",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
